+++ include/model/InhomogeneousMarkovChain.hpp
@@ -73,14 +73,14 @@ class InhomogeneousMarkovChain
   // Trainer
   static SelfPtr train(TrainerPtr<Standard, Self> trainer,
                        wam_algorithm,
-                       unsigned int alphabet_size,
-                       unsigned int order,
-                       unsigned int length,
-                       unsigned int offset,
-                       unsigned int vicinity_length,
+                       size_t alphabet_size,
+                       size_t order,
+                       size_t length,
+                       size_t offset,
+                       size_t vicinity_length,
                        double pseudo_counts,
                        Sequence fixed_sequence,
-                       unsigned int fixed_sequence_position,
+                       size_t fixed_sequence_position,
                        std::vector<double> weights);
 
   /*==========================[ OVERRIDEN METHODS ]===========================*/
@@ -88,39 +88,39 @@ class InhomogeneousMarkovChain
 
   // StandardEvaluator
   Probability evaluateSymbol(SEPtr<Standard> evaluator,
-                             unsigned int pos,
-                             unsigned int phase) const override;
+                             size_t pos,
+                             size_t phase) const override;
   Probability evaluateSequence(SEPtr<Standard> evaluator,
-                               unsigned int begin,
-                               unsigned int end,
-                               unsigned int phase) const override;
+                               size_t begin,
+                               size_t end,
+                               size_t phase) const override;
 
   // CachedEvaluator
   void initializeCache(CEPtr<Standard> evaluator,
-                       unsigned int phase) override;
+                       size_t phase) override;
   Probability evaluateSymbol(CEPtr<Standard> evaluator,
-                             unsigned int pos,
-                             unsigned int phase) const override;
+                             size_t pos,
+                             size_t phase) const override;
   Probability evaluateSequence(CEPtr<Standard> evaluator,
-                               unsigned int begin,
-                               unsigned int end,
-                               unsigned int phase) const override;
+                               size_t begin,
+                               size_t end,
+                               size_t phase) const override;
 
   // StandardGenerator
   Standard<Symbol> drawSymbol(SGPtr<Standard> generator,
-                              unsigned int pos,
-                              unsigned int phase,
+                              size_t pos,
+                              size_t phase,
                               const Sequence& context) const override;
   Standard<Sequence> drawSequence(SGPtr<Standard> generator,
-                                  unsigned int size,
-                                  unsigned int phase) const override;
+                                  size_t size,
+                                  size_t phase) const override;
 
   // SimpleSerializer
   void serialize(SSPtr serializer) override;
 
   /*===========================[ VIRTUAL METHODS ]============================*/
 
-  virtual unsigned int maximumTimeValue();
+  virtual size_t maximumTimeValue();
 
  protected:
   // Instance variables
