+++ include/model/PeriodicInhomogeneousMarkovChain.hpp
@@ -76,9 +76,9 @@ class PeriodicInhomogeneousMarkovChain
   // Trainer
   static SelfPtr train(TrainerPtr<Standard, Self> trainer,
                        interpolation_algorithm,
-                       unsigned int alphabet_size,
-                       unsigned int order,
-                       unsigned int nphases,
+                       size_t alphabet_size,
+                       size_t order,
+                       size_t nphases,
                        double pseudo_counts,
                        std::vector<double> weights,
                        ProbabilisticModelPtr apriori);
@@ -88,32 +88,32 @@ class PeriodicInhomogeneousMarkovChain
 
   // StandardEvaluator
   Probability evaluateSymbol(SEPtr<Standard> evaluator,
-                             unsigned int pos,
-                             unsigned int phase) const override;
+                             size_t pos,
+                             size_t phase) const override;
   Probability evaluateSequence(SEPtr<Standard> evaluator,
-                               unsigned int begin,
-                               unsigned int end,
-                               unsigned int phase) const override;
+                               size_t begin,
+                               size_t end,
+                               size_t phase) const override;
 
   // CachedEvaluator
   void initializeCache(CEPtr<Standard> evaluator,
-                       unsigned int phase) override;
+                       size_t phase) override;
   Probability evaluateSymbol(CEPtr<Standard> evaluator,
-                             unsigned int pos,
-                             unsigned int phase) const override;
+                             size_t pos,
+                             size_t phase) const override;
   Probability evaluateSequence(CEPtr<Standard> evaluator,
-                               unsigned int begin,
-                               unsigned int end,
-                               unsigned int phase) const override;
+                               size_t begin,
+                               size_t end,
+                               size_t phase) const override;
 
   // StandardGenerator
   Standard<Symbol> drawSymbol(SGPtr<Standard> generator,
-                              unsigned int pos,
-                              unsigned int phase,
+                              size_t pos,
+                              size_t phase,
                               const Sequence& context) const override;
   Standard<Sequence> drawSequence(SGPtr<Standard> generator,
-                                  unsigned int size,
-                                  unsigned int phase) const override;
+                                  size_t size,
+                                  size_t phase) const override;
 
   // SimpleSerializer
   void serialize(SSPtr serializer) override;
